Pequenas modificações

+++ humv_backend/src/main/java/br/edu/ufrb/lasis/humv/utils/NumberUtils.java
@@ -1,10 +1,12 @@
 package br.edu.ufrb.lasis.humv.utils;
 
+import java.math.BigInteger;
+
 public class NumberUtils {
 
-	public static Integer convertStringToInteger(String str){
+	public static BigInteger convertStringToInteger(String str){
 		if(str.matches("\\d+")){
-			return Integer.parseInt(str);
+			return BigInteger.valueOf(Integer.parseInt(str));
 		}else{
 			return null;
 		}
